+++ apps/build-onchain-apps/app/buy-me-coffee/index.tsx
@@ -1,9 +1,10 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import BuyMeCoffeeContractDemo from '../../src/components/buy-me-coffee/ContractDemo';
-import Disclaimer from '../../src/components/disclaimer/Disclaimer';
+import Banner from '../../src/components/banner/banner';
 import Header from '../../src/components/header/Header';
+import BuyMeCoffeeContractDemo from '../../src/pages/buy-me-coffee/ContractDemo';
+import Guide from '../../src/pages/buy-me-coffee/Guide';
 
 /**
  * Use the page component to wrap the components
@@ -22,10 +23,10 @@ export default function BuyMeCoffeePage() {
   return (
     <>
       <Header />
-      <main className="container mx-auto flex flex-col px-8">
-        <Disclaimer pageName="Buy Me A Coffee" />
-
+      <main className="container mx-auto flex flex-col px-8 py-6">
+        <Banner pageName="Buy Me A Coffee" pageUrl="buy-me-coffee" />
         <BuyMeCoffeeContractDemo />
+        <Guide />
       </main>
     </>
   );
