+++ apps/build-onchain-apps/app/signature-mint/index.tsx
@@ -1,14 +1,15 @@
 'use client';
 
 import dynamic from 'next/dynamic';
-import Disclaimer from '../../src/components/disclaimer/Disclaimer';
+import Banner from '../../src/components/banner/banner';
 import Header from '../../src/components/header/Header';
+import Guide from '../../src/pages/signature-mint/Guide';
 
 // Because the mint page relies so heavily on client-side state, without disabling SSR
 // for its internals we get annoying hydration errors. A future enhancement would be to
 // read token metadata through a provider that is available server-side.
 const SignatureMint = dynamic(
-  async () => import('../../src/components/mint/SignatureMint').then((mod) => mod),
+  async () => import('../../src/pages/signature-mint/ContractDemo').then((mod) => mod),
   {
     ssr: false,
   },
@@ -22,9 +23,10 @@ export default function MintPage() {
   return (
     <>
       <Header />
-      <main className="container mx-auto flex flex-col">
-        <Disclaimer pageName="Signature Mint" />
+      <main className="container mx-auto flex flex-col px-8 py-6">
+        <Banner pageName="Signature Mint" pageUrl="signature-mint" />
         <SignatureMint />
+        <Guide />
       </main>
     </>
   );
