+++ apps/build-onchain-apps/contracts/src/BuyMeACoffee.sol
@@ -1,6 +1,25 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.23;
-
+/**
+ * ----------------------------------------------------------------------------------------------------------------
+ * ---------██████╗ ██╗   ██╗██╗██╗     ██████╗        ██████╗ ███╗   ██╗ ██████╗██╗  ██╗ █████╗ ██╗███╗   ██╗-----
+ * ---------██╔══██╗██║   ██║██║██║     ██╔══██╗      ██╔═══██╗████╗  ██║██╔════╝██║  ██║██╔══██╗██║████╗  ██║-----
+ * ---------██████╔╝██║   ██║██║██║     ██║  ██║█████╗██║   ██║██╔██╗ ██║██║     ███████║███████║██║██╔██╗ ██║-----
+ * ---------██╔══██╗██║   ██║██║██║     ██║  ██║╚════╝██║   ██║██║╚██╗██║██║     ██╔══██║██╔══██║██║██║╚██╗██║-----
+ * ---------██████╔╝╚██████╔╝██║███████╗██████╔╝      ╚██████╔╝██║ ╚████║╚██████╗██║  ██║██║  ██║██║██║ ╚████║-----
+ * ---------╚═════╝  ╚═════╝ ╚═╝╚══════╝╚═════╝        ╚═════╝ ╚═╝  ╚═══╝ ╚═════╝╚═╝  ╚═╝╚═╝  ╚═╝╚═╝╚═╝  ╚═══╝-----
+ * ----------------------------------------------------------------------------------------------------------------
+ * https://github.com/coinbase/build-onchain-apps
+ *
+ * Disclaimer: The provided Solidity contracts are intended solely for educational purposes and are
+ *   not warranted for any specific use. They have not been audited and may contain vulnerabilities, hence should
+ *   not be deployed in production environments. Users are advised to seek professional review and conduct a
+ *   comprehensive security audit before any real-world application to mitigate risks of financial loss or other
+ *   consequences. The author(s) disclaim all liability for any damages arising from the use of these contracts.
+ *   Use at your own risk, acknowledging the inherent risks of smart contract technology on the blockchain.
+ *
+ */
+ 
 /**
  * @title Memos
  * @dev Memo struct
