+++ apps/build-onchain-apps/src/pages/mint/Guide.tsx
@@ -1,0 +1,10 @@
+export default function Guide() {
+  return (
+    <>
+      <h3 className="mb-6 text-4xl font-medium text-white" id="guide">
+        Guide
+      </h3>
+      <div className="h-px bg-white" />
+    </>
+  );
+}
